some more warnings fixes in the tests

+++ test/e3x_exchange.c
@@ -9,7 +9,7 @@ int main(int argc, char **argv)
   // create A
   lob_t idA = e3x_generate();
   fail_unless(idA);
-  printf("idA key %.*s secret %.*s\n",lob_linked(idA)->head_len,lob_linked(idA)->head,idA->head_len,idA->head);
+  printf("idA key %.*s secret %.*s\n",(int)lob_linked(idA)->head_len,lob_linked(idA)->head,(int)idA->head_len,idA->head);
   e3x_self_t selfA = e3x_self_new(idA,NULL);
   fail_unless(selfA);
   lob_t keyA = lob_get_base32(lob_linked(idA),"1a");
