+++ test/lib_murmur.c
@@ -4,8 +4,8 @@
 int main(int argc, char **argv)
 {
   char hex[9];
-  uint8_t *str = (uint8_t*)"foo bar";
-  uint32_t len = strlen((char*)str);
+  uint32_t *str = (uint32_t*)"foo bar"; // TODO fix this!
+  size_t len = strlen((char*)str);
 
   fail_unless(murmur4(str,len) == 171088983l);
   fail_unless(strcmp("0a329c57",murmur8(str,len,hex)) == 0);
